Move proposals related files and fix notification roles selector (#377)
* move proposals related files

* minor fixes

* fix pre-commit command
