+++ apps/new/widget/page/proposals/NotificationRolesSelector.jsx
@@ -1,6 +1,6 @@
 const DaoSDK = VM.require("sdks.near/widget/SDKs.Sputnik.DaoSDK") || (() => {});
 
-const { InputField } = VM.require("${config_account}/widget/components") || {
+const { InputField } = VM.require("${alias_old}/widget/components") || {
   InputField: <></>,
 };
 
@@ -24,7 +24,9 @@ const group = sdk.getGroupsAndMembers();
 if (group === null || !group.length) {
   return;
 }
-setGroupsAndMembers(group);
+if (!groupsAndMembers.length) {
+  setGroupsAndMembers(group);
+}
 
 const handleCheckboxChange = (role) => {
   setSelectedRoles((prevRoles) => {
@@ -46,9 +48,11 @@ const ThemeContainer =
   props.ThemeContainer ||
   styled.div`
     --primary-color: rgb(255, 175, 81);
+    --text-color: white;
   `;
 
 const Wrapper = styled.div`
+  color: var(--text-color) !important;
   .checked > span:first-child {
     background: var(--primary-color) !important;
     border-color: var(--primary-color) !important;
@@ -64,6 +68,11 @@ const Wrapper = styled.div`
 
   label {
     font-size: 13px;
+    margin-bottom: 5px;
+  }
+
+  .text {
+    color: var(--text-color) !important;
   }
 `;
 
@@ -98,7 +107,8 @@ const createNotificationsData = () => {
                   page: "proposal",
                 },
                 type: "buildhub/custom",
-                widget: "${config_account}/widget/home",
+                widget:
+                  "${config_account}/widget/Index?page=activity&tab=proposals",
               },
             };
           }),
@@ -134,12 +144,13 @@ const groupList = useMemo(() => {
         return null;
       }
       return (
-        <div key={group}>
+        <div key={group.name}>
           <Widget
+            loading=""
             src="nearui.near/widget/Input.Checkbox"
             props={{
               label: (
-                <div>
+                <div className="text">
                   {capitalizeFirstLetter(group.name)} ({membersLength} members)
                 </div>
               ),
@@ -159,13 +170,14 @@ return (
       <div>Send notification to following roles: (Optional)</div>
       <div>
         <label htmlFor={"notifications-label" + daoId}>Custom Message</label>
-        <input
+        <textarea
           name={"notifications-label" + daoId}
           id={"notifications-label" + daoId}
           className="form-control"
           data-bs-theme={bootstrapTheme}
           value={message}
           onChange={(e) => setMessage(e.target.value)}
+          rows="2"
         />
       </div>
       <div>{groupList}</div>
