+++ apps/new/widget/page/proposals/create/FunctionCall.jsx
@@ -1,30 +1,26 @@
-const { Button } = VM.require("${config_account}/widget/components") || {
+const { Button } = VM.require("${alias_old}/widget/components") || {
   Button: () => <></>,
 };
 const { ProposalVisibilityInfoModal } = VM.require(
-  "${config_account}/widget/components.modals.propose.ProposalVisibilityInfoModal",
+  "${config_account}/widget/page.proposals.VisibilityInfoModal",
 ) || {
   ProposalVisibilityInfoModal: () => <></>,
 };
-const DaoSDK =
-  VM.require("sdks.near/widget/SDKs.Sputnik.DaoSDK") || (() => <></>);
+const DaoSDK = VM.require("sdks.near/widget/SDKs.Sputnik.DaoSDK") || (() => {});
 
-const [accountId, setAccountId] = useState("");
-const [role, setRole] = useState("");
-const [isInfoModalActive, setInfoModalActive] = useState(false);
-
-const roles = props.roles;
-const selectedDAO = props.selectedDAO;
-
-const sdk = DaoSDK(selectedDAO);
-if (!sdk) {
-  return <></>;
-}
-
+const [contract, setContract] = useState("");
+const [method, setMethod] = useState("");
+const [args, setArgs] = useState("{}");
+const [gas, setGas] = useState(180000000000000);
+const [deposit, setDeposit] = useState(0);
+const [validatedAddresss, setValidatedAddress] = useState(true);
 const [text, setText] = useState("");
 const [editorKey, setEditorKey] = useState(0);
+const [notificationsData, setNotificationData] = useState(null);
+const [isInfoModalActive, setInfoModalActive] = useState(false);
 
 const bootstrapTheme = props.bootstrapTheme;
+
 useEffect(() => {
   if (!props.item) {
     return;
@@ -34,17 +30,20 @@ useEffect(() => {
   setEditorKey((editorKey) => editorKey + 1);
 }, [props.item]);
 const memoizedKey = useMemo((editorKey) => editorKey, [editorKey]);
-const [validatedAddresss, setValidatedAddresss] = useState(true);
-const [notificationsData, setNotificationData] = useState(null);
+const selectedDAO = props.selectedDAO;
+const sdk = DaoSDK(selectedDAO);
+if (!sdk) {
+  return <></>;
+}
 
 const regex = /.{1}\.near$/;
 useEffect(() => {
-  if (regex.test(accountId) || accountId === "") {
-    setValidatedAddresss(true);
+  if (regex.test(contract) || contract === "") {
+    setValidatedAddress(true);
   } else {
-    setValidatedAddresss(false);
+    setValidatedAddress(false);
   }
-});
+}, [contract]);
 
 const MarkdownEditor = `
   html {
@@ -176,29 +175,48 @@ const TextareaWrapper = styled.div`
 `;
 
 const sdkCall = () => {
-  sdk.createAddMemberProposal({
+  sdk.createFunctionCallProposal({
     description: text,
-    memberId: accountId,
-    roleId: role,
+    receiverId: contract,
+    methodName: method,
+    args: args,
+    proposalDeposit: deposit,
+    proposalGas: gas,
     gas: 180000000000000,
     deposit: 200000000000000,
     additionalCalls: notificationsData,
   });
 };
 
+const NotificationSelector = useMemo(() => {
+  return (
+    <Widget
+      loading=""
+      src="${config_account}/widget/page.proposals.NotificationRolesSelector"
+      props={{
+        daoId: selectedDAO,
+        onUpdate: (v) => {
+          setNotificationData(v);
+        },
+        proposalType: "Function Call",
+      }}
+    />
+  );
+}, [selectedDAO]);
+
 return (
   <div className="d-flex flex-column">
     <div className="form-group mb-3">
-      <label htmlFor="accountId">
-        Account ID<span className="text-danger">*</span>
+      <label htmlFor="contract">
+        Contract<span className="text-danger">*</span>
       </label>
       <input
-        name="accountId"
-        id="accountId"
-        className="form-control"
+        name="contract"
+        id="contract"
         data-bs-theme={bootstrapTheme}
-        value={accountId}
-        onChange={(e) => setAccountId(e.target.value)}
+        value={contract}
+        onChange={(e) => setContract(e.target.value)}
+        className="form-control"
       />
       {!validatedAddresss && (
         <span className="text-danger" style={{ fontSize: 12 }}>
@@ -206,33 +224,63 @@ return (
         </span>
       )}
     </div>
-
     <div className="form-group mb-3">
-      <label htmlFor="role">
-        Role<span className="text-danger">*</span>
+      <label htmlFor="method">
+        Method<span className="text-danger">*</span>
       </label>
-      <select
-        name="role"
-        id="role"
+      <input
+        name="method"
+        id="method"
         data-bs-theme={bootstrapTheme}
-        class="form-select"
-        onChange={(e) => setRole(e.target.value)}
-        selected={role}
-      >
-        <option value="">Select a role</option>
-        {roles.length > 0 &&
-          roles.map((role) => <option value={role}>{role}</option>)}
-      </select>
+        value={method}
+        onChange={(e) => setMethod(e.target.value)}
+        className="form-control"
+      />
     </div>
-
     <div className="form-group mb-3">
-      <label htmlFor="description">Proposal Description</label>
+      <label htmlFor="args">Arguments (JSON)</label>
+      <textarea
+        name="args"
+        id="args"
+        data-bs-theme={bootstrapTheme}
+        value={args}
+        onChange={(e) => setArgs(e.target.value)}
+        className="form-control"
+      />
+    </div>
+    <div className="form-group mb-3">
+      <label htmlFor="gas">Gas</label>
+      <input
+        name="gas"
+        id="gas"
+        type="number"
+        data-bs-theme={bootstrapTheme}
+        value={gas}
+        onChange={(e) => setGas(e.target.value)}
+        className="form-control"
+      />
+    </div>
+    <div className="form-group mb-3">
+      <label htmlFor="deposit">Deposit</label>
+      <input
+        name="deposit"
+        id="deposit"
+        type="number"
+        data-bs-theme={bootstrapTheme}
+        value={deposit}
+        onChange={(e) => setDeposit(e.target.value)}
+        className="form-control"
+      />
+    </div>
+    <div className="form-group mb-3">
+      <label>Proposal Description</label>
       <TextareaWrapper
         className="markdown-editor mb-3"
         data-value={text || ""}
         key={memoizedKey}
       >
         <Widget
+          loading=""
           src="${alias_mob}/widget/MarkdownEditorIframe"
           props={{
             initialText: text,
@@ -244,32 +292,24 @@ return (
         />
       </TextareaWrapper>
     </div>
-    <Widget
-      src="${config_account}/widget/Notification.NotificationRolesSelector"
-      props={{
-        daoId: selectedDAO,
-        onUpdate: (v) => {
-          setNotificationData(v);
-        },
-        proposalType: "Add Member",
-      }}
-    />
+    {NotificationSelector}
+
     <div className="w-100 d-flex">
       <Button
+        disabled={!contract || !method || !validatedAddresss}
         className="ms-auto"
         variant="primary"
-        disabled={!accountId || !role || !validatedAddresss}
         onClick={() => {
           setInfoModalActive(true);
         }}
       >
         Create
       </Button>
+      <ProposalVisibilityInfoModal
+        open={isInfoModalActive}
+        setInfoModalActive={setInfoModalActive}
+        sdkCall={sdkCall}
+      />
     </div>
-    <ProposalVisibilityInfoModal
-      open={isInfoModalActive}
-      setInfoModalActive={setInfoModalActive}
-      sdkCall={sdkCall}
-    />
   </div>
 );
