+++ package.json
@@ -18,10 +18,11 @@
     "start": "yarn dev:gateway",
     "serve": "webpack serve",
     "test": "npx playwright test",
-    "lint": "eslint src --ext js,jsx --report-unused-disable-directives --max-warnings 0"
+    "lint": "eslint src --ext js,jsx --report-unused-disable-directives --max-warnings 0",
+    "prepare": "simple-git-hooks"
   },
   "simple-git-hooks": {
-    "pre-commit": "npm run fmt"
+    "pre-commit": "npx prettier --write '**/*.{js,jsx,ts,tsx,json}'"
   },
   "dependencies": {
     "@keypom/selector": "1.2.3",
